Refactored mod out of public lobby
