+++ src/common/PublicLobby.ts
@@ -1,44 +1,4 @@
 import { GameState } from './AmongUsState';
-export type MODS = "NONE" | "TOWN_OF_IMPOSTORS" | "TOWN_OF_US" | "THE_OTHER_ROLES" | "EXTRA_ROLES" | "POLUS_GG" | "OTHER";
-export interface AmongusMod {
-	id: MODS;
-	label: string;
-	dllStartsWith?: string
-}
-
-
-export const modList: AmongusMod[] = [
-	// recieve this later from git?
-	{
-		id: 'NONE',
-		label: 'None',
-	},
-	{
-		id: 'TOWN_OF_US',
-		label: 'Town of Us',
-		dllStartsWith: 'TownOfUs'
-	},
-	{
-		id: 'TOWN_OF_IMPOSTORS',
-		label: 'Town of Impostors',
-		dllStartsWith: 'TownOfImpostors'
-
-	},
-	{
-		id: 'THE_OTHER_ROLES',
-		label: 'The Other Roles',
-		dllStartsWith: 'TheOtherRoles'
-	},
-	{
-		id: 'POLUS_GG',
-		label: 'Polus.gg',
-		dllStartsWith: 'Polus'
-	},
-	{
-		id: 'OTHER',
-		label: 'Other',
-	},
-];
 
 export interface PublicLobby {
 	id: number;
